requirements.txt
streamlit==1.31.0
google-generativeai==0.8.4
yt-dlp==2023.12.30
PyPDF2==3.0.1
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
Pillow==10.2.0
opencv-python==4.9.0.80
